changed move function to use ints (math with floats is bad practice)

+++ enemy.cpp
@@ -21,11 +21,12 @@ public:
     {
 
     }
+    
     void Draw() {
         DrawTexture(texture, position.x, position.y, WHITE);
     }
 
-    void Move(float deltaX, float deltaY) {
+    void Move(int deltaX, int deltaY) {
         setPosition({position.x + deltaX, position.y + deltaY});
     }
 };
